\name{initJudging}
\alias{initJudging}
\title{
Initialize the judging table.
}
\description{
This reads the judges, sessions and students tables, and
makes an initial guess at the judging table (the link from student
to judge) based on both being in the same session.
}
\usage{
initJudging()
}

\value{
Called for the side effect of creating the judging table.
}

