\name{replaceTable}
\alias{replaceTable}
\title{
Replace a table with a data frame
}
\description{
For debugging and repair:  replace without
checking!
}
\usage{
replaceTable(table, df)
}
\arguments{
  \item{table}{
The name of a table
}
  \item{df}{
A data frame
}
}

\value{
Generally TRUE if it works.
}
