\name{updateStudents}
\alias{updateStudents}
\title{
Update data in the Students table.
}
\description{
Sets the values for students based on a dataframe of new values,
keyed on \code{idnum}.
}
\usage{
updateStudents(db)
}
\arguments{
  \item{db}{
A dataframe containing \code{idnum} and some other fields
from the Students table.
}
}

\value{
No useful value.
}
