\name{showCompetitors}
\alias{showCompetitors}
\title{
Produce a report on all competitors.
}
\description{
}
\usage{
showCompetitors(output = "~/work/SSC/StudentAwards/Student Presentation Awards/Competition 2017/competitors.pdf")
}
\arguments{
  \item{output}{
The desired name of the PDF output file.
}
}
\details{
Writes an R Markdown file to produce the report on all
competitors.  Currently does not render it.
}
\value{
Called for the side effect of producing the report.
}
