\name{showTable}
\alias{showTable}
\title{
Shows a table for debugging purposes.
}
\description{
Shows a table for debugging purposes.
}
\usage{
showTable(name)
}
\arguments{
  \item{name}{
The name of the table to show.
}
}

\value{
A data frame containing the table.
}

\examples{
showTable("Students")
}
